+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeleOp.kt
@@ -9,6 +9,7 @@ import org.firstinspires.ftc.teamcode.modular.BaseLinearOpMode
 import org.firstinspires.ftc.teamcode.modular.GamepadButton
 import org.firstinspires.ftc.teamcode.modular.GamepadState
 import org.firstinspires.ftc.teamcode.modular.ToggleableState
+import org.firstinspires.ftc.teamcode.modular.toggleDirection
 import java.util.Locale
 import kotlin.math.abs
 import kotlin.math.sign
@@ -36,8 +37,16 @@ class TeleOp : BaseLinearOpMode() {
             GamepadButton(gp1, Gamepad::left_bumper) to power::left,
             GamepadButton(gp1, Gamepad::right_bumper) to power::right,
             GamepadButton(gp2, Gamepad::dpad_right) to { bucket.position = 0.0 },
-            GamepadButton(gp2, Gamepad::dpad_down) to { ratchet.engage(); manualRatchetEngagement = true; arm.power = 0.0 },
-            GamepadButton(gp2, Gamepad::dpad_up) to { ratchet.disengage(); manualRatchetEngagement = false; arm.power = 0.0 },
+
+            GamepadButton(gp2, Gamepad::dpad_down) to { ratchet.engage()
+                                                        manualRatchetEngagement = true
+                                                        arm.power = 0.0 },
+
+            GamepadButton(gp2, Gamepad::dpad_up) to { ratchet.disengage()
+                                                      manualRatchetEngagement = false
+                                                      arm.power = 0.0 },
+
+            GamepadButton(gp1, Gamepad::x) to {allMotors.forEach {it.toggleDirection()}}
         )
 
         // TODO: try-catch this to print any errors / force stop the program?
@@ -47,6 +56,7 @@ class TeleOp : BaseLinearOpMode() {
         this.waitForStart()
 
         var lastSwitch = switch.isPressed
+        ratchet.disengage()
 
         while (this.opModeIsActive()) {
 
@@ -98,19 +108,17 @@ class TeleOp : BaseLinearOpMode() {
                 arm.mode = DcMotor.RunMode.STOP_AND_RESET_ENCODER
                 arm.mode = DcMotor.RunMode.RUN_USING_ENCODER
                 ratchet.engage()
-                Thread.sleep(500)
                 bucket.position = 0.225
             }
 
             // going up while arm is at bottom
             if (!manualRatchetEngagement && ratchet.engaged && (-gp2.current.right_stick_y > 0 && currSwitch || -gp2.current.right_stick_y < 0 && !currSwitch /* should never happen */)) {
                 ratchet.disengage()
-                Thread.sleep(500)
             }
 
             when {
                 // block if ratchet is engaged
-                ratchet.engaged -> {}
+                ratchet.engaged -> { arm.power = 0.0 }
 
                 // arm going up
                 -gp2.current.right_stick_y > 0 && arm.currentPosition < 7000 -> {
@@ -121,10 +129,6 @@ class TeleOp : BaseLinearOpMode() {
                 -gp2.current.right_stick_y < 0 && !currSwitch -> {
                     bucket.position = 0.4
                     arm.power = -gp2.current.right_stick_y * 3.0 / 4
-                }
-
-                else -> {
-                    arm.power = 0.0
                 }
             }
