+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/modular/Ratchet.kt
@@ -1,25 +1,29 @@
 package org.firstinspires.ftc.teamcode.modular
 
 import com.qualcomm.robotcore.hardware.Servo
-import java.lang.Thread.sleep
 
-class Ratchet(private val servo: Servo) {
+// active ~ closed
+class Ratchet(val servo: Servo, val active: Double, val inactive: Double) {
+    private var toggled = false;
 
-    // Sleeping is to give the ratchet a chance to (dis)engage
-    private var engaged = false // Don't know if the ratchet is not engaged at start. I hope not lol
-
-    fun engage() {
-        servo.position = 0.075
-        sleep(300)
-        engaged = true
+    fun toggle() {
+        toggled = !toggled
+        servo.position = if (toggled) active else inactive
+        Thread.sleep(500)
     }
 
-   fun disengage() {
-        servo.position = 0.0
-        sleep(300)
-        engaged = false
-   }
+    val engaged
+        get() = toggled
 
-    val isEngaged: Boolean
-        get() = engaged
+    fun engage() {
+        toggled = true
+        servo.position = active
+        Thread.sleep(500)
+    }
+
+    fun disengage() {
+        toggled = false;
+        servo.position = inactive
+        Thread.sleep(500)
+    }
 }
